posx 8 posxben 8.5, need tp know why

+++ srcs/parsing/cub3d.h
@@ -100,8 +100,8 @@ typedef	struct s_dda
 {
 	int		mapx;
 	int		mapy;
-	int		stepx;
-	int		stepy;
+	double	stepx;
+	double	stepy;
 	int		side;
 	int		hit;
 }				t_dda;
